feat(db): setup db and backend
Creates a Postgres database and a web service backend express server,
both running on Render. The frontend is hosted on Netlify, at
https://glistening-vacherin-45f36f.netlify.app/

There is no auth, and there is no switching over to localStorage.

DatabaseStorageStrategy is used by the backend server to send SQL to
the database. RemoteStorageStrategy is used by the front end to send
requests and fetch from the database.

The tasks table in the database stores dates as type DATE, which is a
string. However, the fetched JavaScript is converted into a Date object
unless TO_CHAR is used in the query.