refactor(TaskStats): convert TaskStats class into namespace
The TaskStats class has only static methods, so it can be converted to a
namespace (i.e. a TaskStats object) with static methods as member
functions.

The Set-checking functions (getProjects, getContexts and getPriorities)
can simply return the array containing the set. It is unnecessary to
`return projects || []` as projects will always be an array, empty or
otherwise.